Added comma-separated split to subNumbers()
Rather than splitting every character, I created a regex (uh oh) to insert a comma after every non-digit, or every digit where the next character is a non-digit. The regex will read floating points too, no worries.

This also turned your traditional for-loop into a for-each loop, which saves you a whOLE 32 BYTES OF MEMORY EVERY ITERATION. I got you. As far as I have tested, it is functionally identical to what you had before; let me know if this isn't the case.
e.g:
12345*3.5+5a   ->   12345,*,3.5,+,5,a

The regex can even be modified if you wish to, for example, include commas after groups of [a-z] characters too, so you can have things like trigonometric functions without the need for formatString() in ShuntingYard; all you'll need is .toLowerCase(). This approach also makes the whole thing easier to understand e.g. if you was debugging, you wouldn't need to check what letter corresponds to what functions, because sometimes it isn't very intuitive (e.g. cot -> "u" ???).
I'll leave that decision to you.

If you're curious on the pseudo-regex:
Match any character that isn't a digit or a decimal, which also is not the last character of the string (otherwise a comma will be inserted at the end of the string).
Or, match a digit character where the next character is not a digit or a decimal.